backend/requirements/base.txt
# Core Django packages
Django>=5.0,<5.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.5

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# Security
argon2-cffi>=23.1.0
django-ratelimit>=4.1.0

# Environment variables
python-decouple>=3.8

# Utils
Pillow>=10.2.0
pytz>=2024.1

# API Documentation
drf-spectacular>=0.27.0
